etl/requirements.txt
Flask>=2.0.0,<3.0.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-bigquery-storage>=2.0.0
pandas>=1.0.0
pandas-gbq>=0.26.1
numpy>=1.18.0
h3>=3.7.6
gunicorn>=20.1.0
db-dtypes>=1.1.1
